+++ theseus/geometry/functional/__init__.py
@@ -3,4 +3,4 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-import theseus.geometry.functions.so3_function as SO3Function
+import theseus.geometry.functional.lie_group as LieGroup
